Player
El jugador cuando deja de interactuar regenera la estamina
El jugador tiene invulnerabilidad cuando esquiva

+++ GothicSouls/Assets/Scripts/Enemies/EnemyStats.cs
@@ -19,7 +19,7 @@ namespace SG
             currentHealth = maxHealth;
         }
 
-        private int SetMaxHealthFromHealthLevel()
+        private float SetMaxHealthFromHealthLevel()
         {
             maxHealth = healthLevel * 10;
             return maxHealth;
