+++ GothicSouls/Assets/Scripts/Managers/CharacterStats.cs
@@ -7,12 +7,12 @@ namespace SG
     public class CharacterStats : MonoBehaviour
     {
         public int healthLevel = 10;
-        public int maxHealth;
-        public int currentHealth;
+        public float maxHealth;
+        public float currentHealth;
 
         public int staminaLevel = 10;
-        public int maxStamina;
-        public int currentStamina;
+        public float maxStamina;
+        public float currentStamina;
 
         public bool isDead;
     }
